fix icedfrisby.js - hostname resolves, no response
Hostnames that resolve but do not respond create false positives within
icedfrisby. In a nutshell tests pass when they shouldn't.

It appears there is left over code specific to frisby and jasmine that set the
test as failed before calling done(). Mocha doesn't care and we can probably
remove or modify that code. (it.request_.failedRequest = 1)

In order to prevent this we've added an argument to after() within the toss()
function that will check for an error and call done(). If an arg is set we call
done with Error(), if not done is called normally. This fixes issues with the
bug and limits after() to getting only called once within the codebase with a
specific error and argument.